+++ example/storybook/.storybook/preview.js
@@ -1,0 +1,52 @@
+import { useEffect, Suspense } from 'react';
+import { I18nextProvider } from 'react-i18next';
+
+import i18n from '../src/i18n';
+
+// Create a global variable called locale in storybook
+// and add a dropdown in the toolbar to change your locale
+export const globalTypes = {
+  locale: {
+    name: 'Locale',
+    description: 'Internationalization locale',
+    toolbar: {
+      icon: 'globe',
+      items: [
+        { value: 'en', right: '🇺🇸', title: 'English' },
+        { value: 'de', right: '🇩🇪', title: 'Deutsch' },
+        { value: 'ar', right: '🇦🇪', title: 'عربي' },
+      ],
+      showName: true,
+    },
+  },
+};
+
+// When The language changes, set the document direction
+i18n.on('languageChanged', (locale) => {
+  const direction = i18n.dir(locale);
+  document.dir = direction;
+});
+
+// Wrap your stories in the I18nextProvider component
+const i18nextStoryDecorator = (Story, context) => {
+  const { locale } = context.globals;
+
+  // When the locale global changes
+  // Set the new locale in i18n
+  useEffect(() => {
+    i18n.changeLanguage(locale);
+  }, [locale]);
+
+  return (
+    // here catches the suspense from components not yet ready (still loading translations)
+    // alternative set useSuspense false on i18next.options.react when initializing i18next
+    <Suspense fallback={<div>loading translations...</div>}>
+      <I18nextProvider i18n={i18n}>
+        <Story />
+      </I18nextProvider>
+    </Suspense>
+  );
+};
+
+// export decorators for storybook to wrap your stories in
+export const decorators = [i18nextStoryDecorator];