+++ example/test-jest/src/UseTranslationWithInterpolation.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+
+export default function CustomComponent() {
+  const { t } = useTranslation();
+
+  return <div>{t('some.key', { some: 'variable' })}</div>;
+}