+++ react-i18next.js
@@ -2,9 +2,7 @@
   typeof exports === 'object' && typeof module !== 'undefined' ? factory(exports, require('react')) :
   typeof define === 'function' && define.amd ? define(['exports', 'react'], factory) :
   (global = global || self, factory(global.ReactI18next = {}, global.React));
-}(this, function (exports, React) { 'use strict';
-
-  var React__default = 'default' in React ? React['default'] : React;
+}(this, function (exports, react) { 'use strict';
 
   function ownKeys(object, enumerableOnly) {
     var keys = Object.keys(object);
@@ -320,9 +318,8 @@
     }
   };
 
-  var replace = ''.replace;
-  var es = /&(?:amp|#38|lt|#60|gt|#62|apos|#39|quot|#34);/g;
-  var unes = {
+  var matchHtmlEntity = /&(?:amp|#38|lt|#60|gt|#62|apos|#39|quot|#34|nbsp|#160|copy|#169|reg|#174|hellip|#8230|#x2F|#47);/g;
+  var htmlEntities = {
     '&amp;': '&',
     '&#38;': '&',
     '&lt;': '<',
@@ -332,15 +329,26 @@
     '&apos;': "'",
     '&#39;': "'",
     '&quot;': '"',
-    '&#34;': '"'
+    '&#34;': '"',
+    '&nbsp;': ' ',
+    '&#160;': ' ',
+    '&copy;': '©',
+    '&#169;': '©',
+    '&reg;': '®',
+    '&#174;': '®',
+    '&hellip;': '…',
+    '&#8230;': '…',
+    '&#x2F;': '/',
+    '&#47;': '/'
   };
-  function unescape(un) {
-    return replace.call(un, es, cape);
-  }
 
-  function cape(m) {
-    return unes[m];
-  }
+  var unescapeHtmlEntity = function unescapeHtmlEntity(m) {
+    return htmlEntities[m];
+  };
+
+  var unescape = function unescape(text) {
+    return text.replace(matchHtmlEntity, unescapeHtmlEntity);
+  };
 
   var defaultOptions = {
     bindI18n: 'languageChanged',
@@ -349,10 +357,11 @@
     transSupportBasicHtmlNodes: true,
     transWrapTextNodes: '',
     transKeepBasicHtmlNodesFor: ['br', 'strong', 'i', 'p'],
-    useSuspense: true
+    useSuspense: true,
+    unescape: unescape
   };
   var i18nInstance;
-  var I18nContext = React__default.createContext();
+  var I18nContext = react.createContext();
   function setDefaults() {
     var options = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
     defaultOptions = _objectSpread2(_objectSpread2({}, defaultOptions), options);
@@ -524,13 +533,13 @@
 
   function getChildren(node) {
     if (!node) return [];
-    return node && node.children ? node.children : node.props && node.props.children;
+    return node.props ? node.props.children : node.children;
   }
 
   function hasValidReactChildren(children) {
     if (Object.prototype.toString.call(children) !== '[object Array]') return false;
     return children.every(function (child) {
-      return React__default.isValidElement(child);
+      return react.isValidElement(child);
     });
   }
 
@@ -553,7 +562,7 @@
     childrenArray.forEach(function (child, childIndex) {
       if (typeof child === 'string') {
         stringNode += "".concat(child);
-      } else if (React__default.isValidElement(child)) {
+      } else if (react.isValidElement(child)) {
         var childPropsCount = Object.keys(child.props).length;
         var shouldKeepChild = keepArray.indexOf(child.type) > -1;
         var childChildren = child.props.children;
@@ -602,7 +611,7 @@
       var childrenArray = getAsArray(childs);
       childrenArray.forEach(function (child) {
         if (typeof child === 'string') return;
-        if (hasChildren(child)) getData(getChildren(child));else if (_typeof(child) === 'object' && !React__default.isValidElement(child)) Object.assign(data, child);
+        if (hasChildren(child)) getData(getChildren(child));else if (_typeof(child) === 'object' && !react.isValidElement(child)) Object.assign(data, child);
       });
     }
 
@@ -619,7 +628,7 @@
 
     function pushTranslatedJSX(child, inner, mem, i, isVoid) {
       if (child.dummy) child.children = inner;
-      mem.push(React__default.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
+      mem.push(react.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
         key: i
       }), isVoid ? undefined : inner));
     }
@@ -637,7 +646,7 @@
           var child = Object.keys(node.attrs).length !== 0 ? mergeProps({
             props: node.attrs
           }, tmp) : tmp;
-          var isElement = React__default.isValidElement(child);
+          var isElement = react.isValidElement(child);
           var isValidTranslationWithChildren = isElement && hasChildren(node, true) && !node.voidElement;
           var isEmptyTransWithHTML = emptyChildrenButNeedsHandling && _typeof(child) === 'object' && child.dummy && !isElement;
           var isKnownComponent = _typeof(children) === 'object' && children !== null && Object.hasOwnProperty.call(children, node.name);
@@ -651,7 +660,7 @@
             } else if (isEmptyTransWithHTML) {
             var _inner = mapAST(reactNodes, node.children, rootReactNode);
 
-            mem.push(React__default.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
+            mem.push(react.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
               key: i
             }), _inner));
           } else if (Number.isNaN(parseFloat(node.name))) {
@@ -661,13 +670,13 @@
               pushTranslatedJSX(child, _inner2, mem, i, node.voidElement);
             } else if (i18nOptions.transSupportBasicHtmlNodes && keepArray.indexOf(node.name) > -1) {
               if (node.voidElement) {
-                mem.push(React__default.createElement(node.name, {
+                mem.push(react.createElement(node.name, {
                   key: "".concat(node.name, "-").concat(i)
                 }));
               } else {
                 var _inner3 = mapAST(reactNodes, node.children, rootReactNode);
 
-                mem.push(React__default.createElement(node.name, {
+                mem.push(react.createElement(node.name, {
                   key: "".concat(node.name, "-").concat(i)
                 }, _inner3));
               }
@@ -682,21 +691,21 @@
             var content = node.children[0] ? translationContent : null;
             if (content) mem.push(content);
           } else if (node.children.length === 1 && translationContent) {
-            mem.push(React__default.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
+            mem.push(react.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
               key: i
             }), translationContent));
           } else {
-            mem.push(React__default.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
+            mem.push(react.cloneElement(child, _objectSpread2(_objectSpread2({}, child.props), {}, {
               key: i
             })));
           }
         } else if (node.type === 'text') {
           var wrapTextNodes = i18nOptions.transWrapTextNodes;
 
-          var _content = shouldUnescape ? unescape(i18n.services.interpolator.interpolate(node.content, opts, i18n.language)) : i18n.services.interpolator.interpolate(node.content, opts, i18n.language);
+          var _content = shouldUnescape ? i18nOptions.unescape(i18n.services.interpolator.interpolate(node.content, opts, i18n.language)) : i18n.services.interpolator.interpolate(node.content, opts, i18n.language);
 
           if (wrapTextNodes) {
-            mem.push(React__default.createElement(wrapTextNodes, {
+            mem.push(react.createElement(wrapTextNodes, {
               key: "".concat(node.name, "-").concat(i)
             }, _content));
           } else {
@@ -732,7 +741,7 @@
         shouldUnescape = _ref.shouldUnescape,
         additionalProps = _objectWithoutProperties(_ref, _excluded2);
 
-    var _ref2 = React.useContext(I18nContext) || {},
+    var _ref2 = react.useContext(I18nContext) || {},
         i18nFromContext = _ref2.i18n,
         defaultNSFromContext = _ref2.defaultNS;
 
@@ -773,14 +782,22 @@
     var translation = key ? t(key, combinedTOpts) : defaultValue;
     var content = renderNodes(components || children, translation, i18n, reactI18nextOptions, combinedTOpts, shouldUnescape);
     var useAsParent = parent !== undefined ? parent : reactI18nextOptions.defaultTransParent;
-    return useAsParent ? React__default.createElement(useAsParent, additionalProps, content) : content;
+    return useAsParent ? react.createElement(useAsParent, additionalProps, content) : content;
   }
+
+  var usePrevious = function usePrevious(value, ignore) {
+    var ref = react.useRef();
+    react.useEffect(function () {
+      ref.current = ignore ? ref.current : value;
+    }, [value, ignore]);
+    return ref.current;
+  };
 
   function useTranslation(ns) {
     var props = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
     var i18nFromProps = props.i18n;
 
-    var _ref = React.useContext(I18nContext) || {},
+    var _ref = react.useContext(I18nContext) || {},
         i18nFromContext = _ref.i18n,
         defaultNSFromContext = _ref.defaultNS;
 
@@ -818,13 +835,15 @@
       return i18n.getFixedT(null, i18nOptions.nsMode === 'fallback' ? namespaces : namespaces[0], keyPrefix);
     }
 
-    var _useState = React.useState(getT),
+    var _useState = react.useState(getT),
         _useState2 = _slicedToArray(_useState, 2),
         t = _useState2[0],
         setT = _useState2[1];
 
-    var isMounted = React.useRef(true);
-    React.useEffect(function () {
+    var joinedNS = namespaces.join();
+    var previousJoinedNS = usePrevious(joinedNS);
+    var isMounted = react.useRef(true);
+    react.useEffect(function () {
       var bindI18n = i18nOptions.bindI18n,
           bindI18nStore = i18nOptions.bindI18nStore;
       isMounted.current = true;
@@ -833,6 +852,10 @@
         loadNamespaces(i18n, namespaces, function () {
           if (isMounted.current) setT(getT);
         });
+      }
+
+      if (ready && previousJoinedNS && previousJoinedNS !== joinedNS && isMounted.current) {
+        setT(getT);
       }
 
       function boundReset() {
@@ -850,15 +873,15 @@
           return i18n.store.off(e, boundReset);
         });
       };
-    }, [i18n, namespaces.join()]);
-    var isInitial = React.useRef(true);
-    React.useEffect(function () {
+    }, [i18n, joinedNS]);
+    var isInitial = react.useRef(true);
+    react.useEffect(function () {
       if (isMounted.current && !isInitial.current) {
         setT(getT);
       }
 
       isInitial.current = false;
-    }, [i18n]);
+    }, [i18n, keyPrefix]);
     var ret = [t, i18n, ready];
     ret.t = t;
     ret.i18n = i18n;
@@ -880,7 +903,9 @@
         var forwardedRef = _ref.forwardedRef,
             rest = _objectWithoutProperties(_ref, _excluded$1);
 
-        var _useTranslation = useTranslation(ns, rest),
+        var _useTranslation = useTranslation(ns, _objectSpread2(_objectSpread2({}, rest), {}, {
+          keyPrefix: options.keyPrefix
+        })),
             _useTranslation2 = _slicedToArray(_useTranslation, 3),
             t = _useTranslation2[0],
             i18n = _useTranslation2[1],
@@ -898,19 +923,19 @@
           passDownProps.forwardedRef = forwardedRef;
         }
 
-        return React__default.createElement(WrappedComponent, passDownProps);
+        return react.createElement(WrappedComponent, passDownProps);
       }
 
       I18nextWithTranslation.displayName = "withI18nextTranslation(".concat(getDisplayName(WrappedComponent), ")");
       I18nextWithTranslation.WrappedComponent = WrappedComponent;
 
       var forwardRef = function forwardRef(props, ref) {
-        return React__default.createElement(I18nextWithTranslation, Object.assign({}, props, {
+        return react.createElement(I18nextWithTranslation, Object.assign({}, props, {
           forwardedRef: ref
         }));
       };
 
-      return options.withRef ? React__default.forwardRef(forwardRef) : I18nextWithTranslation;
+      return options.withRef ? react.forwardRef(forwardRef) : I18nextWithTranslation;
     };
   }
 
@@ -936,13 +961,13 @@
     var i18n = _ref.i18n,
         defaultNS = _ref.defaultNS,
         children = _ref.children;
-    var value = React.useMemo(function () {
+    var value = react.useMemo(function () {
       return {
         i18n: i18n,
         defaultNS: defaultNS
       };
     }, [i18n, defaultNS]);
-    return React.createElement(I18nContext.Provider, {
+    return react.createElement(I18nContext.Provider, {
       value: value
     }, children);
   }
@@ -951,7 +976,7 @@
     var props = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : {};
     var i18nFromProps = props.i18n;
 
-    var _ref = React.useContext(I18nContext) || {},
+    var _ref = react.useContext(I18nContext) || {},
         i18nFromContext = _ref.i18n;
 
     var i18n = i18nFromProps || i18nFromContext || getI18n();
@@ -984,7 +1009,7 @@
             rest = _objectWithoutProperties(_ref, _excluded$3);
 
         useSSR(initialI18nStore, initialLanguage);
-        return React__default.createElement(WrappedComponent, _objectSpread2({}, rest));
+        return react.createElement(WrappedComponent, _objectSpread2({}, rest));
       }
 
       I18nextWithSSR.getInitialProps = composeInitialProps(WrappedComponent);