+++ src/Trans.js
@@ -1,6 +1,5 @@
-import React, { useContext } from 'react';
+import { useContext, isValidElement, cloneElement, createElement } from 'react';
 import HTML from 'html-parse-stringify';
-import { unescape } from 'html-escaper';
 import { getI18n, I18nContext, getDefaults } from './context';
 import { warn, warnOnce } from './utils';
 
@@ -13,12 +12,12 @@ function hasChildren(node, checkLength) {
 
 function getChildren(node) {
   if (!node) return [];
-  return node && node.children ? node.children : node.props && node.props.children;
+  return node.props ? node.props.children : node.children;
 }
 
 function hasValidReactChildren(children) {
   if (Object.prototype.toString.call(children) !== '[object Array]') return false;
-  return children.every((child) => React.isValidElement(child));
+  return children.every((child) => isValidElement(child));
 }
 
 function getAsArray(data) {
@@ -49,7 +48,7 @@ export function nodesToString(children, i18nOptions) {
       // actual e.g. lorem
       // expected e.g. lorem
       stringNode += `${child}`;
-    } else if (React.isValidElement(child)) {
+    } else if (isValidElement(child)) {
       const childPropsCount = Object.keys(child.props).length;
       const shouldKeepChild = keepArray.indexOf(child.type) > -1;
       const childChildren = child.props.children;
@@ -124,8 +123,7 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts, s
     childrenArray.forEach((child) => {
       if (typeof child === 'string') return;
       if (hasChildren(child)) getData(getChildren(child));
-      else if (typeof child === 'object' && !React.isValidElement(child))
-        Object.assign(data, child);
+      else if (typeof child === 'object' && !isValidElement(child)) Object.assign(data, child);
     });
   }
 
@@ -145,7 +143,7 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts, s
 
   function pushTranslatedJSX(child, inner, mem, i, isVoid) {
     if (child.dummy) child.children = inner; // needed on preact!
-    mem.push(React.cloneElement(child, { ...child.props, key: i }, isVoid ? undefined : inner));
+    mem.push(cloneElement(child, { ...child.props, key: i }, isVoid ? undefined : inner));
   }
 
   // reactNode (the jsx root element or child)
@@ -171,7 +169,7 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts, s
         const child =
           Object.keys(node.attrs).length !== 0 ? mergeProps({ props: node.attrs }, tmp) : tmp;
 
-        const isElement = React.isValidElement(child);
+        const isElement = isValidElement(child);
 
         const isValidTranslationWithChildren =
           isElement && hasChildren(node, true) && !node.voidElement;
@@ -203,14 +201,14 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts, s
             node.children,
             rootReactNode,
           );
-          mem.push(React.cloneElement(child, { ...child.props, key: i }, inner));
+          mem.push(cloneElement(child, { ...child.props, key: i }, inner));
         } else if (Number.isNaN(parseFloat(node.name))) {
           if (isKnownComponent) {
             const inner = renderInner(child, node, rootReactNode);
             pushTranslatedJSX(child, inner, mem, i, node.voidElement);
           } else if (i18nOptions.transSupportBasicHtmlNodes && keepArray.indexOf(node.name) > -1) {
             if (node.voidElement) {
-              mem.push(React.createElement(node.name, { key: `${node.name}-${i}` }));
+              mem.push(createElement(node.name, { key: `${node.name}-${i}` }));
             } else {
               const inner = mapAST(
                 reactNodes /* wrong but we need something */,
@@ -218,7 +216,7 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts, s
                 rootReactNode,
               );
 
-              mem.push(React.createElement(node.name, { key: `${node.name}-${i}` }, inner));
+              mem.push(createElement(node.name, { key: `${node.name}-${i}` }, inner));
             }
           } else if (node.voidElement) {
             mem.push(`<${node.name} />`);
@@ -242,17 +240,19 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts, s
         } else if (node.children.length === 1 && translationContent) {
           // If component does not have children, but translation - has
           // with this in component could be components={[<span class='make-beautiful'/>]} and in translation - 'some text <0>some highlighted message</0>'
-          mem.push(React.cloneElement(child, { ...child.props, key: i }, translationContent));
+          mem.push(cloneElement(child, { ...child.props, key: i }, translationContent));
         } else {
-          mem.push(React.cloneElement(child, { ...child.props, key: i }));
+          mem.push(cloneElement(child, { ...child.props, key: i }));
         }
       } else if (node.type === 'text') {
         const wrapTextNodes = i18nOptions.transWrapTextNodes;
         const content = shouldUnescape
-          ? unescape(i18n.services.interpolator.interpolate(node.content, opts, i18n.language))
+          ? i18nOptions.unescape(
+              i18n.services.interpolator.interpolate(node.content, opts, i18n.language),
+            )
           : i18n.services.interpolator.interpolate(node.content, opts, i18n.language);
         if (wrapTextNodes) {
-          mem.push(React.createElement(wrapTextNodes, { key: `${node.name}-${i}` }, content));
+          mem.push(createElement(wrapTextNodes, { key: `${node.name}-${i}` }, content));
         } else {
           mem.push(content);
         }
@@ -339,5 +339,5 @@ export function Trans({
   // and override `defaultTransParent` if is present
   const useAsParent = parent !== undefined ? parent : reactI18nextOptions.defaultTransParent;
 
-  return useAsParent ? React.createElement(useAsParent, additionalProps, content) : content;
+  return useAsParent ? createElement(useAsParent, additionalProps, content) : content;
 }