+++ src/withTranslation.js
@@ -1,11 +1,11 @@
-import React from 'react';
+import { createElement, forwardRef as forwardRefReact } from 'react';
 import { useTranslation } from './useTranslation';
 import { getDisplayName } from './utils';
 
 export function withTranslation(ns, options = {}) {
   return function Extend(WrappedComponent) {
     function I18nextWithTranslation({ forwardedRef, ...rest }) {
-      const [t, i18n, ready] = useTranslation(ns, rest);
+      const [t, i18n, ready] = useTranslation(ns, { ...rest, keyPrefix: options.keyPrefix });
 
       const passDownProps = {
         ...rest,
@@ -18,7 +18,7 @@ export function withTranslation(ns, options = {}) {
       } else if (!options.withRef && forwardedRef) {
         passDownProps.forwardedRef = forwardedRef;
       }
-      return React.createElement(WrappedComponent, passDownProps);
+      return createElement(WrappedComponent, passDownProps);
     }
 
     I18nextWithTranslation.displayName = `withI18nextTranslation(${getDisplayName(
@@ -28,8 +28,8 @@ export function withTranslation(ns, options = {}) {
     I18nextWithTranslation.WrappedComponent = WrappedComponent;
 
     const forwardRef = (props, ref) =>
-      React.createElement(I18nextWithTranslation, Object.assign({}, props, { forwardedRef: ref }));
+      createElement(I18nextWithTranslation, Object.assign({}, props, { forwardedRef: ref }));
 
-    return options.withRef ? React.forwardRef(forwardRef) : I18nextWithTranslation;
+    return options.withRef ? forwardRefReact(forwardRef) : I18nextWithTranslation;
   };
 }