+++ test/trans.render.spec.js
@@ -636,9 +636,38 @@ describe('trans should allow escaped html', () => {
         <a
           href="/msgs"
         >
-          &lt;&nbsp;&&gt;
+          &lt; &&gt;
         </a>
         .
+      </div>
+    `);
+  });
+});
+
+describe('trans with custom unescape', () => {
+  let orgValue;
+  beforeAll(() => {
+    orgValue = i18n.options.react.unescape;
+    i18n.options.react.unescape = (text) => text.replace('&shy;', '\u00AD');
+  });
+
+  afterAll(() => {
+    i18n.options.react.unescape = orgValue;
+  });
+
+  it('should allow unescape override', () => {
+    const TestComponent = () => (
+      <Trans i18nKey="transTestCustomUnescape" components={[<Link to="/msgs" />]} shouldUnescape />
+    );
+    const { container } = render(<TestComponent />);
+    expect(container.firstChild).toMatchInlineSnapshot(`
+      <div>
+        Text should be passed through custom unescape 
+        <a
+          href="/msgs"
+        >
+          \u00AD
+        </a>
       </div>
     `);
   });